requirements.txt
streamlit==1.31.1
google-generativeai==0.3.2
Pillow==10.2.0
python-dotenv==1.0.1
pandas==2.2.0
numpy==1.26.3
pydantic==2.6.1
markdown==3.5.2
plotly==5.18.0 